+++ SingleExperience/Services/ProductServices/Models/AddNewProductModel.cs
@@ -8,12 +8,12 @@ namespace SingleExperience.Services.ProductServices.Models
     public class AddNewProductModel
     {
         public string Name { get; set; }
-        public float Price { get; set; }
+        public decimal Price { get; set; }
         public string Detail { get; set; }
         public int Amount { get; set; }
         public CategoryEnum CategoryId { get; set; }
         public int Ranking { get; set; }
         public bool Available { get; set; }
-        public float Rating { get; set; }
+        public decimal Rating { get; set; }
     }
 }
