+++ SingleExperience/Services/ProductServices/Models/ProductSelectedModel.cs
@@ -8,10 +8,10 @@ namespace SingleExperience.Services.ProductServices.Model
     public class ProductSelectedModel
     {
         public int ProductId { get; set; }
-        public double Rating { get; set; }
+        public decimal Rating { get; set; }
         public string Name { get; set; }
         public CategoryEnum CategoryId { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Amount { get; set; }
         public string Detail { get; set; }
     }
